+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Next
 - Add new adapter for CSV/JSON/Parquet files stored in S3
 - Add new adapter for scraping data from HTML tables
 - Add support for ``DROP TABLE``
+- Add registry for adapters
 
 Version 1.0.16 - 2022-07-15
 ===========================